Added YouthAvatars and YouthPlayerList entities, strategies and other related objects. Removed SeniorPlayerSeasonGoals and SeniorPlayerSkills and replaced it with SeniorPlayerWeekLog, also added those entities for JuniorPlayers. Added cancellation loop exit on ChppFileTaskManager so it doesn't need to complete the cycle before cancelling improving performance. Started working on improving the SeniorPlayerGrid on FormMain. Removed IReadOnlyRepository and related objects. Some refactoring and reorganizing.

+++ Hyperar.HattrickUltimate.BusinessObjects/Hattrick/Players/Player.cs
@@ -106,7 +106,7 @@ namespace Hyperar.HattrickUltimate.BusinessObjects.Hattrick.Players
         /// Bruised: 0.
         /// Injured: &gt; 0.
         /// </remarks>
-        public sbyte InjuryLevel { get; set; }
+        public int InjuryLevel { get; set; }
 
         /// <summary>
         /// Gets or sets a value indicating whether the player is foreign or not.
